Issue fixed
Rejecting a Promise always throws an error, this is the way a Promise works, but it doesn't look very neat if our code always throws errors on pages it's not supposed to do anything.
Only way is to not reject the promise, this means we can't use catch/finally should it ever be desired to do something specifically when the poller fails.